some modifications for linux

+++ test/src/objects/1D/curves.cpp
@@ -1,3 +1,6 @@
+//std
+#include <cmath>
+
 //canvas
 #include "inc/Scene/Scene.hpp"
 #include "inc/Objects/1D/Curve.hpp"
@@ -19,18 +22,18 @@ namespace examples
 			//curve
 			canvas::objects::Curve* curve = new canvas::objects::Curve;
 			curve->position([r] (float s){ 
-				const float v1 = r * cos(2 * nt * M_PI * s);
-				const float v2 = r * sin(2 * nt * M_PI * s);
+				const float v1 = r * cosf(2 * nt * M_PI * s);
+				const float v2 = r * sinf(2 * nt * M_PI * s);
 				return canvas::vec3(v1, v2, s);
 			});
 			curve->gradient([r] (float s){ 
-				const float v1 = -2 * nt * M_PI * r * sin(2 * nt * M_PI * s);
-				const float v2 = +2 * nt * M_PI * r * cos(2 * nt * M_PI * s);
+				const float v1 = -2 * nt * M_PI * r * sinf(2 * nt * M_PI * s);
+				const float v2 = +2 * nt * M_PI * r * cosf(2 * nt * M_PI * s);
 				return canvas::vec3(v1, v2, 1);
 			});
 			curve->hessian([r] (float s){ 
-				const float v1 = -4 * nt * nt * M_PI * M_PI * r * cos(2 * nt * M_PI * s);
-				const float v2 = -4 * nt * nt * M_PI * M_PI * r * sin(2 * nt * M_PI * s);
+				const float v1 = -4 * nt * nt * M_PI * M_PI * r * cosf(2 * nt * M_PI * s);
+				const float v2 = -4 * nt * nt * M_PI * M_PI * r * sinf(2 * nt * M_PI * s);
 				return canvas::vec3(v1, v2, 0);
 			});
 			curve->mesh(nm);
